tests/wpt/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/source/navigate-child-function.html

<!doctype html>
<meta charset=utf-8>
<title>Set location from a function called from a parent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe src="support/location-set.html"></iframe>
<script>
  onload = function() {
    var fr = document.querySelector("iframe")
    var url = fr.contentDocument.URL
    fr.contentWindow.go()
    fr.onload = function() {
      assert_equals(fr.contentDocument.referrer, url)
      done()
    }
  }
</script>
