third_party/WebKit/LayoutTests/imported/wpt/html/browsers/browsing-the-web/navigating-across-documents/source/navigate-child-function-expected.txt

This is a testharness.js-based test.
FAIL Set location from a function called from a parent assert_equals: expected "file:///.../html/browsers/browsing-the-web/navigating-across-documents/source/support/location-set.html" but got ""
Harness: the test ran to completion.
